fix: restructured components folder
